cows_project/pyproject.toml
[tool.poetry]
name = "cows_project"
version = "0.1.0"
description = "A Django project for managing cows"
authors = ["cichy767 <vegitto767@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
pydantic = "^2.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
